ErfanGeramizadehNaeini/Q45/cloud/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.6.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
celery
dj-rest-auth
django-allauth
drf-spectacular